chore: Update karpenter role name in main.tf
The code changes in `main.tf` update the role name for the karpenter node in the EKS cluster. The `iam_role_name` is changed from "eks-${module.eks.cluster_name}-karpenter-controller-role" to "eks-${module.eks.cluster_name}-karpenter-node-role". This ensures that the correct role is used for the karpenter node in the EKS cluster.