chore: Add EC2NodeClass for system-addons
This commit adds an EC2NodeClass resource named "system-addons" to the karpenter.yaml file. The EC2NodeClass specifies the AMI family, role, subnet selector terms, security group selector terms, and tags for the system-addons node class. This change is necessary to configure and manage the system-addons node class in the EKS cluster.

Co-authored-by: Carlos Santana <csantana23@gmail.com>